Feat(Product): 상품 price 필드 Long -> Integer로 변경

+++ src/main/java/com/pts/api/product/dto/response/GetProductResponseDto.java
@@ -6,8 +6,8 @@ public record GetProductResponseDto(
     Long id,
     String title,
     String imageUrl,
-    Long minBuyPrice,
-    Long maxSellPrice,
+    Integer minBuyPrice,
+    Integer maxSellPrice,
     LocalDateTime createdAt,
     LocalDateTime updatedAt
 ) {
